CREATE NEW JNLP FILES FOR NEW APPLICATIONS
I have new applications in TEALsim\java\src\tealsim\physics\em and I
have created new prototype jnlp files for each one in
 TEALsim\TEALsim\java\jnlp.

These "prototype" files are picked up by the build file and used to
build a real jnlp file, with the real web address replacing the
"@WebURL@" address in the prototype file.  "@WebURL@" is set in the
build file.

In the process of the build, The real jnlp file is stored in

TEALsim\release\TEALsim-v0.4\jnlp